# web/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MagicStrike API
  description: |
    API para análise tática inteligente de replays de CS2 (Counter-Strike 2).
    
    ### Fluxo de Autenticação (Passwordless)
    1. O usuário solicita um Magic Link através de `/api/v1/auth/magic-link`.
    2. Um e-mail contendo um token seguro de utilização única é enviado (em ambiente de desenvolvimento/teste, o token é registrado nos logs/banco).
    3. O usuário realiza a requisição POST `/api/v1/auth/callback` com o token recebido.
    4. O servidor responde com um JWT (`access_token`) válido por 7 dias e um `session_id`.
    5. O frontend deve incluir o cabeçalho `Authorization: Bearer <access_token>` em todas as rotas protegidas.
    6. Se o usuário deslogar via `DELETE /api/v1/auth/session`, o token de sessão é revogado imediatamente e inserido em uma blocklist em memória com TTL correspondente à expiração do JWT.
    
    ### Upload de Demos (Replays)
    1. O frontend envia metadados da partida e o nome do arquivo para `/api/v1/demos/upload-request`.
    2. O backend responde com uma URL pré-assinada temporária para o MinIO/S3 (`upload_url`) e a chave do bucket correspondente (`bucket_key`).
    3. O frontend faz uma chamada HTTP PUT direta para a `upload_url` enviando o arquivo `.dem` binário (`Content-Type: application/octet-stream`).
...
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de Desenvolvimento Local
paths:
...
  /demos/upload-request:
    post:
      summary: Solicitar URL pré-assinada para Upload de Demo
      description: Registra uma intenção de envio de demo de partida e retorna uma URL segura pré-assinada para upload direto no bucket S3.
      operationId: requestUpload
      tags:
        - Demos & Upload
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - filename
                - team_a
                - team_b
                - map_name
              properties:
                match_id:
                  type: string
                  description: Opcional. ID de identificação customizado se já gerado.
                  example: 01KTAP1DS10RT0VRSBD62Q0MQB
                filename:
                  type: string
                  description: Nome original do arquivo da demo com extensão .dem.
                  example: mibr-vs-lynn-vision-m1-anubis-p2.dem
                md5_hash:
                  type: string
                  description: Hash MD5 opcional do arquivo para verificar duplicatas.
                  example: d41d8cd98f00b204e9800998ecf8427e
                team_a:
                  type: string
                  description: Nome do time A.
                  example: MIBR
                team_b:
                  type: string
                  description: Nome do time B.
                  example: Lynn Vision
                map_name:
                  type: string
                  description: Nome do mapa jogado.
                  example: de_anubis
      responses:
        '200':
          description: URL pré-assinada gerada com sucesso.
          content:
            application/json:
              schema:
                type: object
                required:
                  - upload_url
                  - bucket_key
                  - expires_at
                  - match_id
                properties:
                  upload_url:
                    type: string
                    format: uri
                    description: URL temporária para enviar o arquivo binário via requisição PUT.
                    example: http://127.0.0.1:9002/magicstrike-demos/uploads/01KTAM9NZ3QF5V2N9H62R05JM7/01KTAP1DS10RT0VRSBD62Q0MQB/mibr-vs-lynn-vision-m1-anubis-p2.dem?X-Amz-Algorithm=AWS4-HMAC-SHA256&...
                  bucket_key:
                    type: string
                    description: Caminho identificador único do arquivo no bucket S3. Deve ser passado no payload de confirmação de upload.
                    example: /uploads/01KTAM9NZ3QF5V2N9H62R05JM7/01KTAP1DS10RT0VRSBD62Q0MQB/mibr-vs-lynn-vision-m1-anubis-p2.dem
                  expires_at:
                    type: string
                    format: date-time
                    description: Data de expiração da URL pré-assinada.
                    example: "2026-06-05T01:38:24Z"
                  match_id:
                    type: string
                    description: ID interno associado a esta partida.
                    example: 01KTAP1DS10RT0VRSBD62Q0MQB
        '400':
          description: Requisição inválida ou dados de mapa ausentes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseRFC7807'
        '401':
          description: Não autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseRFC7807'
        '409':
          description: Conflito. Demo duplicada (mesmo MD5 detectado).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseRFC7807'
  /demos/upload-confirm:
    post:
      summary: Confirmar Upload e Enfileirar Processamento
      description: Confirma a presença física do arquivo de demo no bucket S3 e cria uma tarefa assíncrona (job) de parsing no RabbitMQ.
      operationId: confirmUpload
      tags:
        - Demos & Upload
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - match_id
                - bucket_key
              properties:
                match_id:
                  type: string
                  example: 01KTAP1DS10RT0VRSBD62Q0MQB
                bucket_key:
                  type: string
                  example: /uploads/01KTAM9NZ3QF5V2N9H62R05JM7/01KTAP1DS10RT0VRSBD62Q0MQB/mibr-vs-lynn-vision-m1-anubis-p2.dem
      responses:
        '201':
          description: Upload confirmado. Partida adicionada na fila de processamento.
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - match_id
                properties:
                  status:
                    type: string
                    example: queued
                  match_id:
                    type: string
                    example: 01KTAP1DS10RT0VRSBD62Q0MQB
        '400':
          description: Dados em formato incorreto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseRFC7807'
        '401':
          description: Autenticação necessário.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseRFC7807'
        '404':
          description: Partida correspondente ou arquivo no S3 não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseRFC7807'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponseRFC7807:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          example: "about:blank"
        title:
          type: string
          example: "Upload request failed"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "invalid upload request: empty filename"
        instance:
          type: string
          example: "/api/v1/demos/upload-request"